+++ src/db/postgres/sql/subscription-delete-expired.sql
@@ -1,0 +1,4 @@
+--
+DELETE FROM subscription
+WHERE topic_id = $(topicId) AND expires < now()
+